pyproject.toml
[project]
name = "django-query-assistant-panel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "django-debug-toolbar>=5.2.0",
    "google-genai>=1.15.0",
    "google-generativeai>=0.8.5",
    "markdown>=3.8",
]
